Add allowRequestsOnFailure configuration parameter to RateLimiter
Added the allowRequestsOnFailure parameter to the rate limiter In order to allow the user of this library to configure whether the rate limiter will allow or deny requests when an exception occurs. This configuration parameter will give us the flexibility to choose whether we allow or block requests in failure scenarios.

Although there is currently only one implementation of each data store that stores the data in memory, we may eventually want to utilize a Redis distributed cache.  This is one such case where failures outside of our system may cause the rate limiter to fail.